refactor: use Schema from metaschema library
