+++ lib/model-db.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const fs = require('fs').promises;
 
-const schema = require('./schema-fs.js');
+const schema = require('./schema.js');
 const { cpdir, shorten } = require('./utils.js');
 
 const dbms = {};
