+++ lib/model-pg.js
@@ -3,7 +3,7 @@
 const { Client } = require('pg');
 
 const { dbms } = require('./model-db.js');
-const schema = require('./schema-fs.js');
+const schema = require('./schema.js');
 
 const { toLowerCamel, toUpperCamel, isUpperCamel } = require('./utils.js');
